pyproject.toml
[tool.poetry]
name = "electronics-model"
version = "0.1.0"
description = ""
authors = ["Besshtannikova Ekaterina <dolmatova3010@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.7"
dotenv = "^0.9.9"
psycopg2 = "^2.9.10"
drf-spectacular = "^0.28.0"
djangorestframework-simplejwt = "^5.5.0"
pillow = "^11.1.0"
django-filter = "^25.1"
drf-spectacular-sidecar = "^2025.3.1"
django-cors-headers = "^4.7.0"


[tool.poetry.group.lint.dependencies]
flake8 = "^7.1.2"
mypy = "^1.15.0"
black = "^25.1.0"
isort = "^6.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
# максимальная длина строки
line_length = 119

[tool.black]
# Максимальная длина строки
Line-length = 119

[tool.mypy]
disallow_untyped_defs = true
warn_return_any = true
